requirements.txt
asgiref==3.2.7
click==7.1.2
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
gunicorn==20.0.4
image==1.5.31
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
redis==3.5.2
rq==1.4.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
